LWDEV-3468: get accounts by margin usage level, close positions by accounts

+++ src/MarginTrading.Common/ClientContracts/MarginTradingAccountClientContract.cs
@@ -23,7 +23,7 @@ namespace MarginTrading.Common.ClientContracts
         public decimal UsedMargin { get; set; }
         public decimal MarginInit { get; set; }
         public decimal PnL { get; set; }
-        public decimal OpenPositionsCount { get; set; }
+        public int OpenPositionsCount { get; set; }
         public decimal MarginUsageLevel { get; set; }
         public bool IsLive { get; set; }
     }
